move send monitoring sections
